Updated toolchain Added multiple compression methods Fixed bug where not all images would get detected Moved some graphics functions to guiUtils.cpp Fixed loadingBarHandler from getting included multiple times
