+++ src/globals.h
@@ -10,7 +10,7 @@
 * ticalc.org
 */
 #define OBTAINED_FROM " from bin folder"
-#define VERSION "3.0.0-alpha.1" OBTAINED_FROM
+#define VERSION "3.0.0-alpha.2" OBTAINED_FROM
 #define YEAR "2025"
 #define TUTORIAL_LINK "no tutorial for alpha"
